Contained Rana Processing Preparation
Initial refactoring and framework setup for running rana processing
code within a containerized environment to ease the pain of managing
software dependencies will also increasing reproducibility.

Additional scripts were added to help simplify the processing steps.
As an example, a script was added to deal exclusively with manual
pollinator extraction for the purpose of dataset building rather than
relying on a set of flags in the main pollinator_extraction script.

Signed-off-by: Avery Uslaner <avery.uslaner@redbutte.utah.edu>