Manual Pollinator Extraction Script Overhaul
Refactored manual pollinator extraction script to separate elements into
their own functions.

Added filtering of videos to search for frames that have not yet been
processed based on the last entry in the logging database.

Increased verbosity of print statements to make it easier for the user
to understand what was being saved and where.

Signed-off-by: Avery Uslaner <avery.uslaner@redbutte.utah.edu>