Break after None Frame
AttributeErrors as a result of None frames now break the execution loop.
If an AttributeError happens as a result of something other than a None
frame, a message is printed and the continues for another try.

Also changed two related if statements examing frame_time to an if else
clause.

Signed-off-by: Avery Uslaner <avery.uslaner@redbutte.utah.edu>