Flexibility, API Updates, And Eval Metric Additions
Dataset averages file as well as their associated test record files are
now supplied to the test script as commandline arguments.

Model loading now uses the updated Module.load function since the
previously used FeedForward method is deprecated.

The F1 metric was added to the model evaluations as it also provides
access to a record of the number of false positives and negatives when
testing binary classifiers.

Signed-off-by: Avery Uslaner <avery.uslaner@redbutte.utah.edu>